Update and rename 0267. Missing Number.py to 0268. Missing Number.py

+++ array/0268. Missing Number.py
@@ -2,4 +2,4 @@
 
 def missingNumber(self, nums):
     n = len(nums)
-    return n * (n+1) / 2 - sum(nums)
+    return int(n * (n+1) / 2 - sum(nums))
